[Serialization] Attempt to load impl-only dependencies of modules with testing
Implementation-only dependencies may be referenced from internal decls
in a module that enabled testing. If that module is imported as testable
these clients may need to load transitive dependencies.

It would be project breaking to make those dependencies as required so
let's do the next best thing and attempt to load them.

rdar://79459263