IRGen: correct field counting for distributed actors
Correct the counting between `forEachField` and `getNumFields` to
correctly enumerate the synthesized fields.  We would previously fail to
count the non-root default actor resulting in a mismatch on the stored
field counts.  Fortunately, we were able to catch this by means of an
assertion in the compiler which triggered on Windows.

Special thanks to @etcwilde for the help in tracking this down!

rdar://106822386