Remove this over-eager assertion about transforming subst SILFunctionType.
It was firing on mapping a SILFunctionType from pack to element context.
I could've updated that, but honestly it's just not a very good assertion.
The most likely source of errors here is already prevented by the
assertion about substituting a SILFunctionType with AST-level
substitution.  Otherwise we're just interfering with the implementation
of reasonable AST-level transforms.