Fix the computation of thunk types for functions involving element archetypes
The old code assumed that thunk signatures would only possibly refer to a
single opened existential archetype.  To generalize this to opened element
archetypes, it needs to both support multiple archetypes and, of course,
support the new kind of local archetype.  So here's a bunch of code to
clone element signatures and support multiple local archetypes.  It fell out
to also make this support multiple opened existential archetypes, if we
ever have code patterns that require that.