Bump macOS deployment target for TSan to 10.13 for LLVM only preset
This allows LLVM TSAN tests to pass when using Xcode 14.3.

Long term we want to switch to `LLVM_USE_RUNTIMES` to build compiler-rt,
so we can leverage the value of `SANITIZER_MIN_OSX_VERSION` set in
`cmake_product.py`.
In the meantime, we prefer enforcing the deployment in the only preset
that needs this, since this currently requires leveraging an
intermediate variable (`CLANG_COMPILER_RT_CMAKE_ARGS`) which cannot be
easily augmented in presets mixins.

Addresses rdar://106441106