[debug-info] Remove foot-gun code around how we initialize names of SILDebugVariable.
Specifically instead of lazily returning a name based off of the SIL location,
we just set the decl appropriately when we initialize the instruction. If
someone wants to use a different name, they can create a new instruction. As a
nice benefit, we also eliminate a memcpy since we can just reuse the identifier
from the VarDecl.