[NFC] Remove extra comma inside the macros.
The macros are all used with a semicolon after the macro invocation, so
the comma inside the macro and outside the macro are redundant. This
causes a warning in modern compilers with extra warnings enabled that it
is quite spammy in the output.

Remove the commas inside the macro to avoid the warning. This should not
have any meningful effect in the result.