[cxx-interop] Mark clang-related requests as un-cached.
Pointers from clang appear to be unstable in some capacity. My theory is that clang loads one module, then frees its AST when it's done, or maybe just re-allocates the AST at some point. In any case, we cannot cache requests on clang pointers.

This should fix the flakeyness issue that we've been seeing for a while.