Ban compilers without SwiftCC/SwiftAsyncCC
Building the runtime with a compiler that doesn't understand swift
calling conventions is playing with fire, and not a fun fire because the
Swift compiler will emit calls expecting swift calling conventions,
resulting in an ABI mismatch and lots of pain. This patch stops us from
defining the offending macros in cases where the C/C++ compiler building
the runtime doesn't understand Swift calling conventions, so attempted
usage will result in a hard build failure.

Unfortunately, the runtime-inspection library includes this header, but
is part of the toolchain build, which should be buildable with the host
C/C++ compiler (msvc, gcc, etc...), so I can't outright `#error` the
failing clause or it will cause that library to fail to build.
Sadly, unknown attributes are either handled as soft warnings or ignored
altogether, which is not a great situation when building a broken ABI.
We could use `_Pragma("clang diagnostic error
\"-Wattribute-ignored\"")`, but that only works for clang, where my main
target is gcc. I've gone with not defining the macro in cases where the
compiler doesn't understand the calling conventions so that it is a hard
error at the point of use, and emitting a warning saying that we're not
defining the macro because the compiler doesn't understand the calling
convention to explain why.

I've done the same with the swift async calling convention, removing the
fallback on the Swift CC. Windows added swift async calling convention
support in apple/llvm-project commit
e5198240d982c635129cf1a2ae2f2251747a981d, and apple/swift commit
4f6a054705f15de75c39bdcdaefb4737f4917d7b.