[CanOSSALifetime] Xfer lexical value to callees.
When canonicalizing a lexical lifetime, don't treat applies which
consume a copy of the value as deinit barriers.  Doing so forces another
copy of the def being canonicalized to remain after the apply.

Instead, allow the lifetime to be transferred into the callee.  This is
the same behavior that already exists for lexical lifetimes represented
with the `begin_borrow [lexical]` + `copy_value` instruction sequence.